Constructor now needs a queryselector string instead of a dom element id.
This is probably mostly a cosmetic change, but the mostly unnecessary dependence on jQuery bothers me. If a browser supports WebGL or Canvas, it will also support document.querySelector. jQuery is mostly unnecessary here.
We will need to do something about the _src and the _title elements, that surprised me.